Refactor Schedule entity to change TimeZone type from int8 to float for improved flexibility; update related computations for consistency

+++ src/domain/Schedule.fs
@@ -10,4 +10,4 @@ type Schedule =
       StopTime: TimeOnly option
       Workdays: DayOfWeek Set
       Recursively: bool
-      TimeZone: int8 }
+      TimeZone: float }
